+++ Block/Adminhtml/Synccatalog/Grid.php
@@ -15,12 +15,31 @@ class Grid extends \Magento\Backend\Block\Widget\Grid\Extended
      * @var \Saleslayer\Synccatalog\Model\Synccatalog
      */
     protected $_synccatalog;
-    protected $_systemStore;
+
+        /**
+     * @var \Magento\Cron\Model\Schedule
+     */
+    protected $_cronSchedule;
 
     /**
-     * @param \Magento\Backend\Block\Template\Context $context
-     * @param \Magento\Backend\Helper\Data $backendHelper
-     * @param \Saleslayer\Synccatalog\Model\Synccatalog $synccatalog
+     * @var \Magento\Framework\App\ResourceConnection
+     */
+    protected $_connection;
+
+    /**
+     * @var \Magento\Framework\App\DeploymentConfig
+     */
+    protected $_deploymentConfig;
+
+    protected $tablePrefix = null;
+
+    /**
+     * @param \Magento\Backend\Block\Template\Context                                   $context
+     * @param \Magento\Backend\Helper\Data                                              $backendHelper
+     * @param \Saleslayer\Synccatalog\Model\Synccatalog                                 $synccatalog
+     * @param \Magento\Cron\Model\Schedule                                              $cronSchedule
+     * @param \Magento\Framework\App\DeploymentConfig                                   $deploymentConfig
+     * @param \Magento\Framework\App\ResourceConnection                                 $resourceConnection
      * @param \Saleslayer\Synccatalog\Model\ResourceModel\Synccatalog\CollectionFactory $collectionFactory
      * @param array $data
      */
@@ -28,13 +47,17 @@ class Grid extends \Magento\Backend\Block\Widget\Grid\Extended
         \Magento\Backend\Block\Template\Context $context,
         \Magento\Backend\Helper\Data $backendHelper,
         \Saleslayer\Synccatalog\Model\Synccatalog $synccatalog,
+        \Magento\Cron\Model\Schedule $cronSchedule,
+        \Magento\Framework\App\DeploymentConfig $deploymentConfig,
+        \Magento\Framework\App\ResourceConnection $resourceConnection,
         \Saleslayer\Synccatalog\Model\ResourceModel\Synccatalog\CollectionFactory $collectionFactory,
-        \Magento\Store\Model\System\Store $systemStore,
         array $data = []
     ) {
         $this->_collectionFactory = $collectionFactory;
         $this->_synccatalog = $synccatalog;
-        $this->_systemStore = $systemStore;
+        $this->_cronSchedule = $cronSchedule;
+        $this->_deploymentConfig = $deploymentConfig;
+        $this->_connection = $resourceConnection->getConnection();
         parent::__construct($context, $backendHelper, $data);
     }
 
@@ -59,71 +82,86 @@ class Grid extends \Magento\Backend\Block\Widget\Grid\Extended
     protected function _prepareCollection()
     {
         $collection = $this->_collectionFactory->create();
-        /* @var $collection \Saleslayer\Synccatalog\Model\ResourceModel\Synccatalog\Collection */
-
-        // $file = BP.'/var/log/sl_logs/_debbug_log_saleslayer_test.txt';
-
-        // // $stores_data = $this->_systemStore->getStoreValuesForForm(false, true);
-
-        // // $store_data = $this->_systemStore->getStoreName(2);
-        // // file_put_contents($file, 'store_data: '.print_r($store_data,1)."\n", FILE_APPEND);
-
-        
-        // // $storeCollection = $this->_scopeInterface->_loadStoreCollection();
-        // // file_put_contents($file, 'storeCollection: '.print_r($storeCollection,1)."\n", FILE_APPEND);
-        // // $collectionData = $collection->getData();
-        // // // file_put_contents($file, 'test: '.print_R($collectionData,1)."\n", FILE_APPEND);
-
-        // foreach ($collection as $keyConn => $connector) {
-        //     // file_put_contents($file, 'key: '.$keyConn."\n", FILE_APPEND);
-        //     // file_put_contents($file, 'connector: '.print_r($connector->getData(),1)."\n", FILE_APPEND);
-        //     $connectorData = $connector->getData();
-        //     $store_view_ids = json_decode($connectorData['store_view_ids'],1);
-        //     file_put_contents($file, 'store_view_ids: '.print_r($store_view_ids,1)."\n", FILE_APPEND);
-        //     $store_names = array();
-        //     foreach ($store_view_ids as $store_view_id) {
-        // // file_put_contents($file, 'store_view_id: '.print_r($store_view_id,1)."\n", FILE_APPEND);
-        //         if ($store_view_id == 0){
-        //             $store_name = 'All Store Views';
-        //         }else{
-        //             $store_name = $this->_systemStore->getStoreName($store_view_id);
-        //         }
-        //         array_push($store_names, $store_name);
-        //     }
-        //     // if (count($store_names) > 1){
-        //         $store_names = implode(',', $store_names);
-        //         $connectorData['store_view_id'] = $store_names;
-        //     // }else{
-        //     //     $store_names = $store_names[0];
-        //     // }
-        //     // file_put_contents($file, 'store_names: '.$store_names."\n", FILE_APPEND);
-        //     // // $collection[$keyConn['store_view_ids']] = $store_names;
-        //     // $collection->setStoreViewIds('');
-        //     // $collectionData[$keyConn]['store_view_ids'] = $store_names;
-        //     // $name_test = $this->_scopeConfig->getStore()->getName();
-        //     // file_put_contents($file, 'name_test: '.print_r($name_test,1)."\n", FILE_APPEND);
-        //     // file_put_contents($file, 'col id: '.print_r($col->getId(),1)."\n", FILE_APPEND);
-        //     // file_put_contents($file, 'col index: '.print_r($col->getIndex(),1)."\n", FILE_APPEND);
-        //     // file_put_contents($file, 'store_names: '.print_r($store_names,1)."\n", FILE_APPEND);
-        //     $connector->setStoreViewIds($store_names);
-        //     file_put_contents($file, 'connector data before add: '.print_r($connector->getData(),1)."\n", FILE_APPEND);
-        //     // $connector->setData($connectorData);
-        //     // $collection->updateObject($connector);
-        //     $collection->removeItemByKey($keyConn);
-        //     $collection->addItem($connector);
-        //     // $collection->save();
-        // }
-
-        // foreach ($collection as $keyConn => $connector) {
-        
-
-        // }
-        
-        // file_put_contents($file, 'test data: '.print_r($collection->getData(),1)."\n", FILE_APPEND);
-
         $this->setCollection($collection);
 
         return parent::_prepareCollection();
+    }
+
+        /**
+     * After load collection
+     * 
+     * @return \Magento\Backend\Block\Widget\Grid
+     */
+    protected function _afterLoadCollection()
+    {
+        parent::_afterLoadCollection();
+
+        $collection = $this->getCollection();
+        
+        foreach ($collection as $collection_item) {
+            
+            $like_param = '%"connector_id":"'.$collection_item->getData('connector_id').'"%';
+            
+            $items_processing = $this->_connection->fetchAll(
+                $this->_connection->select()
+                    ->from($this->getTable('saleslayer_synccatalog_syncdata'),
+                    ['total' => new \Zend_Db_Expr('COUNT(*)'), 'item_type'])
+                    ->where("sync_params LIKE '".$like_param."'")
+                    ->group('item_type')
+                   
+            );
+
+            if (!empty($items_processing)){
+                
+                $items_processing_message = $cron_error_message = '';
+                $total_items_processing = 0;
+                
+                $item_types_names = [
+                    'category' => 'categories',
+                    'product' => 'products',
+                    'product_format' => 'product formats',
+                    'product_links' => 'product links',
+                    'product__images' => 'product images',
+
+                ];
+                foreach ($items_processing as $items_processing_by_type){
+
+                    $total_items_processing += $items_processing_by_type['total'];
+
+                    if ($items_processing_message === ''){
+                        $items_processing_message = "\n".$items_processing_by_type['total'].' '.$item_types_names[$items_processing_by_type['item_type']];
+                    }else{
+                        $items_processing_message .= "\n".$items_processing_by_type['total'].' '.$item_types_names[$items_processing_by_type['item_type']];
+                    }
+
+                }
+
+                $collection_item_message = '<b>Connector processing</b>: '.$total_items_processing.' items total.';
+                $collection_item_message .= $items_processing_message;
+
+                $sl_last_syncdata_cron = $this->_connection->fetchRow(
+                    $this->_connection->select()
+                        ->from(
+                           $this->getTable('cron_schedule')
+                        )
+                        ->where("job_code = 'Saleslayer_Synccatalog_Syncdatacron'")
+                        ->where("status NOT IN (?)", ['missed', 'pending'])
+                        ->where('executed_at IS NOT NULL')
+                        ->order('scheduled_at DESC')
+                        ->limit(1)
+                );
+
+                if (is_array($sl_last_syncdata_cron) && $sl_last_syncdata_cron['status'] == 'error'){
+                    $cron_error_message .= "<b>CRON ERROR</b>: ".$sl_last_syncdata_cron['messages']."\n";
+                }
+
+                if ($cron_error_message !== '') $collection_item_message = $cron_error_message.$collection_item_message;
+                $collection_item->setData('connector_status', $collection_item_message);
+            }
+
+        }
+
+        return $this;
     }
 
     /**
@@ -133,15 +171,39 @@ class Grid extends \Magento\Backend\Block\Widget\Grid\Extended
      */
     protected function _prepareColumns()
     {
-        $this->addColumn('id', [
-            'header'    => __('ID'),
-            'index'     => 'id',
-        ]);
+        $this->addColumn(
+            'id', 
+            [
+                'header'    => __('ID'),
+                'index'     => 'id',
+                'header_css_class' => 'grid-header-class grid-col-id-class',
+                'column_css_class' => 'grid-col-id-class'
+            ]
+        );
         
-        $this->addColumn('connector_id', ['header' => __('Connector ID'), 'index' => 'connector_id']);
+        $this->addColumn(
+            'connector_id', 
+            [
+                'header' => __('Connector ID'), 
+                'index' => 'connector_id',
+                'header_css_class' => 'grid-header-class grid-col-conn-class',
+                'column_css_class' => 'grid-col-conn-class'
+            ]
+        );
 
-        // $this->addColumn('store_view_ids', ['header' => __('Store View IDs'), 'index' => 'store_view_ids']);
-        // $this->addColumn('secret_key', ['header' => __('Secret Key'), 'index' => 'secret_key']);
+        $this->addColumn(
+            'connector_status', 
+            [
+                'header' => __('Connector Status'), 
+                'index' => 'connector_status',
+                'type' => 'text',
+                'filter' => false,
+                'sortable' => false,
+                'renderer' => 'Saleslayer\Synccatalog\Block\Adminhtml\Renderer\CustomGridText',
+                'header_css_class' => 'grid-header-class grid-col-status-class',
+                'column_css_class' => 'grid-col-status-class'
+            ]
+        );
         
         $this->addColumn(
             'last_update',
@@ -149,32 +211,76 @@ class Grid extends \Magento\Backend\Block\Widget\Grid\Extended
                 'header' => __('Last Update'),
                 'index' => 'last_update',
                 'type' => 'datetime',
-                'header_css_class' => 'col-date',
-                'column_css_class' => 'col-date'
+                'header_css_class' => 'grid-header-class grid-col-update-class col-date',
+                'column_css_class' => 'grid-col-update-class col-date'
             ]
         );
+            
+        return parent::_prepareColumns();
+    }
 
-        // $this->addColumn(
-        //     'sync_data',
-        //     [
-        //         'header' => __('Synchronization Data'),
-        //         'index' => 'sync_ata',
-        //         'header_css_class' => 'col-sync',
-        //         'column_css_class' => 'col-sync'
-        //     ]
-        // );
+    /**
+     * Prepare layout
+     * 
+     * @return \Magento\Backend\Block\Widget\Grid\Extended
+     */
+    protected function _prepareLayout()
+    {
+        parent::_prepareLayout();
 
-        $test = parent::_prepareColumns();
+        $this->pageConfig->addPageAsset(
+            'Saleslayer_Synccatalog::js/grid-reload.js'
+        );
 
-        // $file = BP.'/_debbug_log_saleslayer_test.txt';
+        $this->pageConfig->addPageAsset(
+            'Saleslayer_Synccatalog::css/grid.css'
+        );
+        
+        return $this;
+    }
 
-        // foreach ($test as $key => $t) {
-        //     // file_put_contents($file, 'test_data: '.print_r($test->getMultipleRows($test),1), FILE_APPEND);
-        //     file_put_contents($file, 'test_data: '.print_r($t,1), FILE_APPEND);
-        // }
+    /**
+     * Function to get table name with prefix.
+     * @param string $tableName               table to search
+     * @return string                         table in database with prefix
+     */
+    private function getTable($tableName){
+        
+        $tableNameReturn = $this->_connection->getTableName($tableName);
 
-        return $test;
-        // return parent::_prepareColumns();
+        if ($this->_connection->isTableExists($tableNameReturn)){
+
+            $tablePrefix = $this->getTablePrefix();
+
+            if ($tablePrefix && strpos($tableNameReturn, $tablePrefix) !== 0) {
+
+                $tableNameReturn = $tablePrefix . $tableNameReturn;
+
+            }
+
+            return $tableNameReturn;
+
+        }
+
+        return null;
+
+    }
+
+    /**
+     * Function to get table prefix
+     * @return string                        configuration table prefix
+     */
+    private function getTablePrefix(){
+
+        if (null === $this->tablePrefix) {
+                
+            $this->tablePrefix = (string)$this->_deploymentConfig->get(
+                \Magento\Framework\Config\ConfigOptionsListConstants::CONFIG_PATH_DB_PREFIX
+            );
+        
+        }
+        
+        return $this->tablePrefix;
     }
 
     /**